src/telegram_client/pyproject.toml
[tool.poetry]
name = "telegram-client"
version = "0.1.0"
description = ""
authors = ["Yaroslav Pisarev <yaricp@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
kafka-python = "^2.0.2"
protobuf = "^5.29.0"
six = "^1.17.0"
pydantic = "^2.9.2"
pydantic-extra-types = "^2.10.0"
pytelegrambotapi = "^4.24.0"
pydantic-settings = "^2.6.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
